initial implementation of compare contracts
